pyproject.toml
[project]
name = "adk-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "adktools>=0.1.0",
    "google-adk>=0.1.0",
    "litellm==1.65.1",
    "pip>=25.0.1",
    "requests>=2.31.0",
    "aiosqlite>=0.18.0",
]